+++ packages/as-xlsx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @noy-db/as-xlsx
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/as-zip@0.1.0
+
 ## 1.0.0
 
 ### Patch Changes