+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @noy-db/cli
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/to-meter@0.1.0
+  - @noy-db/to-probe@0.1.0
+
 ## 1.0.0
 
 ### Patch Changes