+++ packages/in-nextjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @noy-db/in-nextjs
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/in-react@0.1.0
+
 ## 1.0.0
 
 ### Patch Changes