+++ packages/in-nuxt/CHANGELOG.md
@@ -1,3 +1,13 @@
 # Changelog — in-nuxt
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/in-pinia@0.1.0
+  - @noy-db/in-rest@0.1.0
+  - @noy-db/in-vue@0.1.0
+
 ## 0.1.0-pre.1 — Initial pre-release