+++ packages/to-cloudflare-r2/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @noy-db/to-cloudflare-r2
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/to-aws-s3@0.1.0
+
 ## 1.0.0
 
 ### Patch Changes