+++ packages/to-supabase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @noy-db/to-supabase
 
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/to-postgres@0.1.0
+
 ## 1.0.0
 
 ### Patch Changes