+++ playground/cli/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @noy-db/playground-cli
 
-## 0.1.0
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/to-file@0.1.0
+  - @noy-db/to-memory@0.1.0
+
+## 0.1.0-pre.7
 
 ### Patch Changes