+++ showcases/CHANGELOG.md
@@ -1,6 +1,59 @@
 # @noy-db/showcases
 
-## 0.1.0
+## 0.1.0-pre.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @noy-db/hub@0.1.0
+  - @noy-db/as-blob@0.1.0
+  - @noy-db/as-csv@0.1.0
+  - @noy-db/as-json@0.1.0
+  - @noy-db/as-ndjson@0.1.0
+  - @noy-db/as-noydb@0.1.0
+  - @noy-db/as-xlsx@0.1.0
+  - @noy-db/as-xml@0.1.0
+  - @noy-db/as-zip@0.1.0
+  - @noy-db/by-peer@0.1.0
+  - @noy-db/by-tabs@0.1.0
+  - @noy-db/in-ai@0.1.0
+  - @noy-db/in-nextjs@0.1.0
+  - @noy-db/in-pinia@0.1.0
+  - @noy-db/in-react@0.1.0
+  - @noy-db/in-rest@0.1.0
+  - @noy-db/in-solid@0.1.0
+  - @noy-db/in-svelte@0.1.0
+  - @noy-db/in-tanstack-query@0.1.0
+  - @noy-db/in-tanstack-table@0.1.0
+  - @noy-db/in-vue@0.1.0
+  - @noy-db/in-yjs@0.1.0
+  - @noy-db/in-zustand@0.1.0
+  - @noy-db/on-email-otp@0.1.0
+  - @noy-db/on-magic-link@0.1.0
+  - @noy-db/on-oidc@0.1.0
+  - @noy-db/on-pin@0.1.0
+  - @noy-db/on-recovery@0.1.0
+  - @noy-db/on-shamir@0.1.0
+  - @noy-db/on-threat@0.1.0
+  - @noy-db/on-totp@0.1.0
+  - @noy-db/on-webauthn@0.1.0
+  - @noy-db/to-aws-dynamo@0.1.0
+  - @noy-db/to-aws-s3@0.1.0
+  - @noy-db/to-browser-idb@0.1.0
+  - @noy-db/to-browser-local@0.1.0
+  - @noy-db/to-cloudflare-d1@0.1.0
+  - @noy-db/to-cloudflare-r2@0.1.0
+  - @noy-db/to-file@0.1.0
+  - @noy-db/to-memory@0.1.0
+  - @noy-db/to-meter@0.1.0
+  - @noy-db/to-postgres@0.1.0
+  - @noy-db/to-probe@0.1.0
+  - @noy-db/to-sqlite@0.1.0
+  - @noy-db/to-supabase@0.1.0
+  - @noy-db/to-turso@0.1.0
+  - @noy-db/to-webdav@0.1.0
+
+## 0.1.0-pre.7
 
 ### Patch Changes